Fix the code that has been broken by previous renaming
